+++ packages/snap-store-mobx/src/index.ts
@@ -10,7 +10,7 @@ export * from './Search/Stores/';
 export * from './Autocomplete/Stores/';
 export { RecommendationStore } from './Recommendation/RecommendationStore';
 export { CartStore } from './Cart/CartStore';
-export { StorageStore, StorageConfig, StorageType } from './Storage/StorageStore';
+export { StorageStore, StorageConfig, StorageType } from '@athoscommerce/snap-toolbox';
 export { MetaStore } from './Meta/MetaStore';
 export { ProductQuickViewStore } from './ProductQuickView/ProductQuickViewStore';
 export { cloneProductForQuickView } from './ProductQuickView/cloneProductForQuickView';